For Garret
fixed issues
