+++ src/main/java/com/tips/TIPS/lib/mysql-connector-j-8.0.32.jar
